cyberly/requirements.txt
Django==5.2.1
djangorestframework
django-cors-headers
django-filter
mysqlclient
gunicorn
